Add test for updating configuration in ServiceProvider
